// src/types/api.d.ts
export interface components {
  schemas: {
// ...
    MentorReviewViewDto: {
      content?: string;
      createdBy?: string;
      /** Format: date-time */
      createdDate?: string;
      elapsedCreatedDate?: string;
      /** @enum {string} */
      entityStatus?: "ACTIVE" | "BLOCKED" | "DELETED" | "INVALID" | "SEARCHABLE";
      lastModifiedBy?: string;
      /** Format: date-time */
      lastModifiedDate?: string;
      /** Format: date */
      localDate?: string;
      mentorName?: string;
      mentorUuid?: string;
      /** Format: bigdecimal */
      rating?: number;
      userName?: string;
      uuid?: string;
    };
    /** MentorViewDto */
    MentorViewDto: {
      availableSlots?: components["schemas"]["MentorAvailableSlotDto"][];
      createdBy?: string;
      /** Format: date-time */
      createdDate?: string;
      description?: string;
      educations?: components["schemas"]["MentorEducationDto"][];
      elapsedCreatedDate?: string;
      /** @enum {string} */
      entityStatus?: "ACTIVE" | "BLOCKED" | "DELETED" | "INVALID" | "SEARCHABLE";
      experiences?: components["schemas"]["MentorExperienceDto"][];
      hashTags?: string[];
      isBookmarkedByMe?: boolean;
      lastModifiedBy?: string;
      /** Format: date-time */
      lastModifiedDate?: string;
      /** Format: int32 */
      maxDiscountedAblePrice?: number;
      name?: string;
      /** Format: int32 */
      originPrice?: number;
      profileImageUrl?: string;
      recurringSlots?: components["schemas"]["MentorRecurringSlotViewDto"][];
      shortDescription?: string;
      uuid?: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}